requirements.txt
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
biopython==1.81
blosc2==2.0.0
certifi==2023.5.7
charset-normalizer==3.2.0
comm==0.1.3
contourpy==1.1.0
cycler==0.11.0
Cython==0.29.36
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
filelock==3.13.4
fonttools==4.41.0
fsspec==2024.3.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
ipykernel==6.24.0
ipython==8.12.2
jedi==0.18.2
Jinja2==3.1.3
joblib==1.3.1
jupyter_client==8.3.0
jupyter_core==5.3.1
kiwisolver==1.4.4
MarkupSafe==2.1.5
matplotlib==3.7.2
matplotlib-inline==0.1.6
mpmath==1.3.0
msgpack==1.0.5
nest-asyncio==1.5.6
networkx==3.3
numexpr==2.8.4
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
packaging==23.1
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.9.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
pyzmq==25.1.0
requests==2.31.0
scikit-learn==1.3.0
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
stack-data==0.6.2
sympy==1.12
tables==3.8.0
threadpoolctl==3.2.0
torch==2.2.2
tornado==6.3.2
traitlets==5.9.0
triton==2.2.0
typing_extensions==4.11.0
tzdata==2023.3
urllib3==2.0.3
wcwidth==0.2.6
wget==3.2
zipp==3.16.2